ROle impl method redefined
